Move wifi events to infrastructure package
