+++ app/src/main/java/com/alive_n_clickin/waft/application/NearbyBusScanner.java
@@ -4,12 +4,10 @@ import android.content.Context;
 
 import com.alive_n_clickin.waft.infrastructure.WifiBroadcastReceiver;
 import com.alive_n_clickin.waft.infrastructure.WifiHelper;
-import com.alive_n_clickin.waft.util.event.CantSearchForVehiclesEvent;
 import com.alive_n_clickin.waft.util.event.IEvent;
 import com.alive_n_clickin.waft.util.event.IObservable;
 import com.alive_n_clickin.waft.util.event.IObservableHelper;
 import com.alive_n_clickin.waft.util.event.IObserver;
-import com.alive_n_clickin.waft.util.event.NewBusNearbyEvent;
 import com.alive_n_clickin.waft.infrastructure.NewWifiScanAvailableEvent;
 import com.alive_n_clickin.waft.util.event.ObservableHelper;
 import com.alive_n_clickin.waft.infrastructure.WifiStateChangeEvent;
